# Validation: fields will merge: Faster argument comparison 